Rename License Data to License Report
This move is to dramatically reduce confusion on terminology with the object structures. License data was the name of the object and the name of a property in the object.
The license report is a more accurate name of the object that contains all of the metadata about licenses rather than the licenses themselves.
The name of the property called license data remains un-changed as that is a good name for it still.
The routes have been updated from license data to license report to more accurately reflect the name of the stuff being messed with.
Permissions have been updated too.