Create SDK Clients
Create code, license, build process, readme, and structure.
This process is able to build SDKs based on the OpenAPI specs in the specs folder.
The SDK folder needs to be generated locally as it is generated content.
The project is structured as a mono repo with a shared TS Config.