Fix Publish Process
The publish process was not allowing the upload of the unziped files without running the Typia initialization process. This can't be skipped in the current state.
Created NPM Ignore files to select only the desired files.
Updated the entry point to select the bin folder since the bin folder will now be published.
Remove the seperate publish job in the actions.
Actions wording.
Add additional tags to the package.