No Trailing Slashes 😭
Just when I thought I had a good standard, the code generator comes and yanks it out from under me. No more trailing slashes.
This honestly matches the Graph API, don't know how I missed this.
This fixes an issue with Kiota's generator where it makes a double route with the trailing slash instead of a single route. This breaks projects that depend on the Kiota SDK.